# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ct-project-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://ct-project.pp.ua/api/"
        },
        {
            "url": "http://172.19.0.1/api/"
        }
    ],
    "paths": {
...
        "/v1/admin/services": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "68aa901cafb7fcbd8b45f712507ae71b",
                "parameters": [
                    {
                        "name": "lastId",
                        "in": "query",
                        "description": "Min: 0",
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "487d67aa8e30b9271399fdf3dc21984d",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "query",
                        "description": "Only existing Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "max: 255",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "max: 500",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 500
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Only existing User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "regex in format: 999.99",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "Only existing City ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create new service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/admin/services/{id}": {
            "get": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "ba089f9b570285388e4979fb08c81488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show the service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "6d32efc9b8bb1726e307f51056a3e867",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete the service"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Panel"
                ],
                "operationId": "ef81fadcfd718c86f9d2fd26c7565734",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "query",
                        "description": "Only existing Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "max: 255",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "max: 500",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 500
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Only existing User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "regex in format: 999.99",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "Only existing City ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show updated service",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CityIdNameAndSlug": {
                "description": "Show information about city id, name and slug",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrors": {
                "description": "Validation Errors",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "description": "each key describes error message",
                        "properties": {
                            "query or path name": {
                                "type": "array",
                                "items": {}
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The email field must be a valid email address. (and 3 more errors)",
                    "errors": {
                        "email": [
                            "The email field must be a valid email address."
                        ],
                        "password": [
                            "The password field confirmation does not match.",
                            "The password field must contain at least one uppercase and one lowercase letter.",
                            "The password field must contain at least one number."
                        ]
                    }
                }
            },
...
            "Service": {
                "description": "Show information about service",
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer"
                    },
                    "category": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "integer"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "firstName": {
                                "type": "string"
                            },
                            "lastName": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "city": {
                        "$ref": "#/components/schemas/CityIdNameAndSlug"
                    },
                    "createdAt": {
                        "type": "integer"
                    },
                    "updatedAt": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "category": {
                        "id": 4,
                        "title": "Догляд за нігтями",
                        "slug": "doglyad-za-nigtyami"
                    },
                    "title": "Заголовок послуги",
                    "description": "Детальний опис послуги",
                    "user": {
                        "id": 32,
                        "firstName": "Олена",
                        "lastName": "Новікова",
                        "email": "olena.novikova@example.com"
                    },
                    "price": 900.45,
                    "city": {
                        "id": 1,
                        "name": "Київ",
                        "slug": "ukrayina-kiyiv"
                    },
                    "createdAt": "2023-12-02 14:51:18",
                    "updatedAt": "2023-12-02 14:51:18"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}